// src/App.UI/src/app/core/api/generated/model/authenticateResponse.ts
/**
 * HMI.API
 * This document describe the endpoints to use in the UI to connect with the entire CVS system.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Role } from './role';


/**
 * The authenticate response.
 */
export interface AuthenticateResponse { 
    /**
     * Gets or sets the id.
     */
    id?: number;
    /**
     * Gets or sets the username.
     */
    username?: string | null;
    /**
     * Gets or sets the first name.
     */
    firstName?: string | null;
    /**
     * Gets or sets the last name.
     */
    lastName?: string | null;
    role?: Role;
    /**
     * Gets or sets the jwt token.
     */
    jwtToken?: string | null;
    /**
     * Gets or sets the refresh token.
     */
    refreshToken?: string | null;
}