// src/App.UI/src/app/core/api/generated/model/role.ts
/**
 * HMI.API
 * This document describe the endpoints to use in the UI to connect with the entire CVS system.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * The roles a user can have.
 */
export type Role = 'Operator' | 'Maintenance' | 'Commissioning' | 'Admin';

export const Role = {
    Operator: 'Operator' as Role,
    Maintenance: 'Maintenance' as Role,
    Commissioning: 'Commissioning' as Role,
    Admin: 'Admin' as Role
};